Add definição de cargo de staff e correção de bugs

+++ src/controller/commands/remover_log.py
@@ -1,8 +1,11 @@
 import interactions
 from src.repositories.discord_repository import logsrepository
+from src.repositories.discord_repository import roledef
 from bson import ObjectId
 
 repo = logsrepository()
+repo2 = roledef()
+
 
 class remover_logs(interactions.Extension):
     def __init__(self, bot):
@@ -15,7 +18,27 @@ class remover_logs(interactions.Extension):
     
     async def logs(self, ctx:interactions.CommandContext, id:str):
         
-        repo.remove_user(ObjectId(id))
+        await ctx.defer()
+
+        role = ctx.member.roles
+        i = 0
+        
+        while len(role) > i:
+            
+            author = ctx.author
+            logs = repo2.find_role(role[i]) 
+            
+            if logs:
+                
+                repo.remove_user(ObjectId(id))
+       
+        else:
+                embed_else = interactions.Embed()
+
+                embed_else.title = 'Erro ao usar o comando'
+                embed_else.description = f'Você não tem permissão para usar o comando'
+                embed_else.color = int('ff0000', 16)
+                await author.send(embeds=embed_else)     
 
 def setup (bot):
     remover_logs(bot)
